Merchant Account Service (#1)
* feat: generate merchant data model


* feat: refactored merchant


* feat: update merchant account


* feat: fix failing deployment file


* feat: remove invalid docker network ref


* feat: update go mod file


* feat: updates


* feat: add feelguuds development helm chart


* feat: update models


* feat: update database interaction logic


* feat: properly configured connection to merchant component db


* feat: fix all startup issues


* feat: update unit test cases


* added unit test for find account by email operation


* added unit test for find merchant account by id operation


* feat: added utility funcs


* feat: updated unit tests


* feat: add merchant account api


* feat: updated merchant account apis


* fixed all failing unit tests


* feat: fix failing unit test in the authsvc


* removed unecessary folders


* remove dependency on auth service


* update git action test


* fixed ref to faulty script link


* update gitignore file


* feat: update helm charts


* update workflow files


* feat: clean up helm files


* feat: remove update flag


* feat: remove update flag


* feat: remove update flag


* update resource ref in yaml file


* feat: update values file


* feat: update values file


* feat: update values file


* feat: update values file


* feat: remove unecessary resource constraint


* feat: remove unecessary resource constraint


* feat: refactor backend code


* feat: fixed failing unit test

+++ internal/authentication_handler/get_account.go
@@ -11,9 +11,9 @@ import (
 )
 
 type GetAccountResponse struct {
-	Account *core_auth_sdk.Account `json:"account"`
-	Code  int    `json:"code"`
-	ErrorMessage string `json:"message"`
+	Account      *core_auth_sdk.Account `json:"account"`
+	Code         int                    `json:"code"`
+	ErrorMessage string                 `json:"message"`
 }
 
 // GetAccountHandler godoc
@@ -44,7 +44,7 @@ func (c *AuthenticationComponent) GetAccountHandler(w http.ResponseWriter, r *ht
 	}
 
 	// invoke authentication service
-	account, err = c.GetAccount(ctx, Id)
+	account, err = c.GetAccount(ctx, uint32(Id))
 	if err != nil {
 		helper.ErrorResponse(w, err.Error(), http.StatusBadRequest)
 		return
@@ -52,7 +52,6 @@ func (c *AuthenticationComponent) GetAccountHandler(w http.ResponseWriter, r *ht
 
 	helper.JSONResponse(w, GetAccountResponse{Account: account})
 }
-
 
 // GetAccount obtains a user account from the context of the authentications service (authn) based on a provided user id
 func (c *AuthenticationComponent) GetAccount(ctx context.Context, Id uint32) (*core_auth_sdk.Account, error) {
